refactor(frontend): 자료요약 폴링 유틸을 공용 위치로 옮긴다
폴링 유틸이 Documents 페이지 안에 있어 다른 화면에서 쓰려면 페이지 폴더를
가로질러 import 해야 했다. 자료요약 말고도 백그라운드 작업 결과를 기다리는
화면이 생기므로 공용 api 폴더로 옮긴다.

- pages/Documents/summaryPolling.ts -> api/polling.ts 로 이동
- DocumentDrawer, useDocuments, summaryPolling 테스트의 import 경로 갱신
